removal of makefile; update to cog validation
