+++ scripts/tiff-to-cog-conversions/tif-to-cog.sbatch.sh
@@ -12,10 +12,20 @@ module load physics gdal/3.5.2
 module load python/3.9.0
 module load py-gdal-utils
 
-rm tif_paths.txt # make sure we are always making a fresh file
-
 # update this path to reflect folders to search through to locate .tif files
 WORKDIR=$OAK/natcap-data-catalog-cache/footprint-impact-tool-data/
+N_WORKERS=8
+VALIDATION=/share/software/user/open/py-gdal-utils/3.4.1_py39/lib/python3.9/site-packages/osgeo_utils/samples/validate_cloud_optimized_geotiff.py
 
+# create list of geotiffs
+python3 list-geotiffs.py $WORKDIR > tif_paths.txt
 
-make WORKDIR=$WORKDIR cog-translate
+for t in $(cat tif_paths.txt); 
+    do
+		j=`echo $t | cut -d . -f 1`; \
+        j=$j"_cog.tif"; \
+        echo $t $j; \
+		GDAL_CACHEMAX=2048 gdal_translate ${t} ${j} -of cog -co "BIGTIFF=YES" -co "NUM_THREADS=$N_WORKERS"; \
+        GDAL_CACHEMAX=2048 python3 $VALIDATION $j >> $WORKDIR/validation_check.txt; \
+		python3 scale-check.py $j $t >> $WORKDIR/cog_scale_check.txt; \
+	done
